+++ lua/kissedit/init.lua
@@ -1,0 +1,15 @@
+local mq = require("mq")
+
+--local drawWindow = require("kissedit")
+
+--local window = drawWindow.DrawMainWindow()
+
+
+
+mq.cmd('/lua run kissedit/kissedit')
+
+
+
+
+
+